# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/moderation/pending:
    get:
      operationId: ReportsController_getPending
      parameters:
      - name: page
        required: false
        in: query
        schema:
          type: number
      - name: limit
        required: false
        in: query
        schema:
          type: number
      responses:
        '200':
          description: Pending reports
      summary: List pending reports (admin)
      tags:
      - Reports
  /admin/moderation/reports/{id}:
    patch:
      operationId: ReportsController_review
      parameters:
      - name: id
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewReportDto'
      responses:
        '200':
          description: Report reviewed
      summary: Review a report (admin)
      tags:
      - Reports
...
  /recommendations:
    get:
      operationId: RecommendationsController_getRecommendations
      parameters:
      - name: limit
        required: false
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecommendationResultDto'
      summary: Get personalized trek recommendations
      tags:
      - Recommendations
  /recommendations/preferences:
    put:
      operationId: RecommendationsController_setPreferences
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecommendationPreferenceDto'
      responses:
        '200':
          description: ''
      summary: Set recommendation preferences
      tags:
      - Recommendations
    get:
      operationId: RecommendationsController_getPreferences
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      summary: Get recommendation preferences
      tags:
      - Recommendations
...
  /reports:
    post:
      operationId: ReportsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReportDto'
      responses:
        '200':
          description: Report created
      summary: Report inappropriate content
      tags:
      - Reports
...
  /treks/{trekId}/itinerary:
    get:
      operationId: ItinerariesController_getByTrek
      parameters:
      - name: trekId
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Itinerary days ordered by day number
      summary: Get full itinerary for a trek
      tags:
      - Itineraries
    put:
      operationId: ItinerariesController_upsertDays
      parameters:
      - name: trekId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        '200':
          description: Updated itinerary days
      summary: Bulk set/reorder all itinerary days
      tags:
      - Itineraries
  /treks/{trekId}/itinerary/days:
    post:
      operationId: ItinerariesController_addDay
      parameters:
      - name: trekId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItineraryDayDto'
      responses:
        '200':
          description: Created itinerary day
      summary: Add a single itinerary day
      tags:
      - Itineraries
  /treks/{trekId}/itinerary/days/{dayId}:
    patch:
      operationId: ItinerariesController_updateDay
      parameters:
      - name: dayId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItineraryDayDto'
      responses:
        '200':
          description: Updated itinerary day
      summary: Update a single itinerary day
      tags:
      - Itineraries
    delete:
      operationId: ItinerariesController_deleteDay
      parameters:
      - name: dayId
        required: true
        in: path
        schema:
          type: string
      responses:
        '200':
          description: Day removed
      summary: Remove an itinerary day
      tags:
      - Itineraries
  /treks/{trekId}/itinerary/reorder:
    patch:
      operationId: ItinerariesController_reorder
      parameters:
      - name: trekId
        required: true
        in: path
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderItineraryDto'
      responses:
        '200':
          description: Reordered itinerary days
      summary: Reorder itinerary days
      tags:
      - Itineraries
...
  /treks/{trekId}/recommendations:
    get:
      operationId: RecommendationsController_getSimilarTreks
      parameters:
      - name: trekId
        required: true
        in: path
        schema:
          type: string
      - name: limit
        required: false
        in: query
        schema:
          type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SimilarTrekDto'
      summary: Get similar treks for a specific trek
      tags:
      - Recommendations
...
info:
  title: Offbeat प्रवासी API
  description: Full-scale, production-ready trek discovery and booking platform.
  version: 1.0.0
  contact: {}
...
servers:
- url: http://localhost:4000
  description: Local development
- url: https://api.offbeatpravasi.com
  description: Production
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
    ApiKey:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
...
    CreateItineraryDayDto:
      type: object
      properties:
        dayNumber:
          type: number
          minimum: 1
        title:
          type: string
          maxLength: 255
        description:
          type: string
        distanceKm:
          type: number
        altitudeGainM:
          type: number
        altitudeLossM:
          type: number
        maxAltitudeM:
          type: number
        mealPlan:
          $ref: '#/components/schemas/MealPlanDto'
        accommodationType:
          type: string
          enum:
          - TENT
          - GUESTHOUSE
          - CAMP
          - HOMESTAY
        activityType:
          type: string
          enum:
          - TREKKING
          - REST
          - ACCLIMATIZATION
          - SIGHTSEEING
      required:
      - dayNumber
      - title
      - activityType
...
    CreateReportDto:
      type: object
      properties:
        targetType:
          enum:
          - POST
          - COMMENT
          - USER
          - TREK
          - REVIEW
          - STORY
          type: string
        targetId:
          type: string
          format: uuid
        reason:
          enum:
          - SPAM
          - INAPPROPRIATE
          - HARASSMENT
          - HATE_SPEECH
          - MISINFORMATION
          - IMPERSONATION
          - VIOLENCE
          - OTHER
          type: string
        description:
          type: string
      required:
      - targetType
      - targetId
      - reason
...
    MealPlanDto:
      type: object
      properties:
        breakfast:
          type: string
        lunch:
          type: string
        dinner:
          type: string
...
    RecommendationPreferenceDto:
      type: object
      properties:
        preferredDifficulty:
          items:
            type: string
            enum:
            - EASY
            - MODERATE
            - DIFFICULT
            - EXTREME
          description: Preferred difficulty levels
          example:
          - MODERATE
          - DIFFICULT
          type: array
        preferredStates:
          description: Preferred states
          example:
          - Uttarakhand
          - Himachal Pradesh
          type: array
          items:
            type: string
        maxBudget:
          type: number
          minimum: 0
          description: Max budget in INR
          example: 15000
        preferredDurationDays:
          maxItems: 2
          description: Preferred duration range [min, max] in days
          example:
          - 3
          - 7
          type: array
          items:
            type: number
            minimum: 0
        interests:
          description: Interest tag IDs
          example:
          - tag-id-1
          - tag-id-2
          type: array
          items:
            type: string
    RecommendationResultDto:
      type: object
      properties:
        trekId:
          type: string
        score:
          type: number
        reason:
          type: string
      required:
      - trekId
      - score
      - reason
...
    ReorderItineraryDto:
      type: object
      properties:
        dayIds:
          type: array
          items:
            type: string
            format: uuid
      required:
      - dayIds
...
    ReviewReportDto:
      type: object
      properties:
        status:
          type: string
          enum:
          - REVIEWED
          - DISMISSED
          - ACTION_TAKEN
        adminNotes:
          type: string
      required:
      - status
...
    SimilarTrekDto:
      type: object
      properties:
        trekId:
          type: string
        score:
          type: number
        reason:
          type: string
      required:
      - trekId
      - score
      - reason
...
    UpdateItineraryDayDto:
      type: object
      properties:
        dayNumber:
          type: number
          minimum: 1
        title:
          type: string
          maxLength: 255
        description:
          type: string
        distanceKm:
          type: number
        altitudeGainM:
          type: number
        altitudeLossM:
          type: number
        maxAltitudeM:
          type: number
        mealPlan:
          $ref: '#/components/schemas/MealPlanDto'
        accommodationType:
          enum:
          - TENT
          - GUESTHOUSE
          - CAMP
          - HOMESTAY
          type: string
        activityType:
          enum:
          - TREKKING
          - REST
          - ACCLIMATIZATION
          - SIGHTSEEING
          type: string
...
security:
- ApiKey: []